### 2022-01-17 #62 @thomasMisiek
[Re] A general model of hippocampal and dorsal striatal learning and decision making
**Original article:** Geerts, J. P., Chersi, F., Stachenfeld, K. L., & Burgess, N. (2020). A general model of hippocampal and dorsal striatal learning and decision making. In Proceedings of the National Academy of Sciences (Vol. 117, Issue 49, pp. 31427–31437). Proceedings of the National Academy of Sciences. https://doi.org/10.1073/pnas.2007981117

**Scientific domain: **Computational Neuroscience, Spatial Navigation
**Programming language:** Python, Jupyter Notebook
