### 2022-01-17 #62 @vitay
Yes.

> In can take up to one week to be computed on a laptop (i5 2.30ghz), if the RTDP algorithm is used as the HPC model, and roughly 2 days if the SR is selected.

But do not expect my review for tomorrow, then...
