### 2022-01-18 #62 @thomasMisiek
Thank you very much for your time ! @vitay if that may help you, only the SR was used in this work, the RTDP model was used for another project.
