### 2022-01-18 #62 @mstimberg
@rougier Thanks for the invite, I'd be happy to review the submission. But just to make sure that the effort is well spent: is it a problem that the original article already comes with reproducible [code on ModelDB](https://senselab.med.yale.edu/ModelDB/showmodel.cshtml?model=266836)? I vaguely remember discussions that submissions for papers with published and accessible code are only appropriate for ReScience C if the replication is a "clean room" implementation, which is clearly not the case here, given the statement in the [README](https://github.com/thomasMisiek/mixed-coordination-models/blob/main/README.md):
> Numerous classes, functions and methods in this work are partially or fully taken from Geerts, Chersi, Stachenfeld and Burgess (2020) original code.

I did not find anything about this in the ReScience FAQ and it is of course possible that I remember this discussion incorrectly.
