### 2022-01-18 #62 @rougier
@mstimberg Oh thank for noticing, I did not check it.
@thomasMisiek Can you explain what is your personal work exactly (and if this is a replication or a reproduction)?
