### 2022-01-18 #62 @vitay
A good deal of the pdf is about that. In short, the code provided by Geerts et al does not allow to reproduce the results of their paper without changing the parameters, nor does it allow to reproduce the original experiments of Pearce et al without changing some critical part of the code. I have not looked in detail yet what the contribution is in terms of lines of code, but functionally this makes sense to me. It is actually both a reproduction of Geerts' paper and a replication of Pearce's experiment (or the opposite?).
I think their approach is valid, but yes, we should discuss it beforehand.
