### 2022-01-18 #62 @vitay
A note on installation: with an empty environment (only python, pip and wheel), pip-installing requirements.txt fails as some dependencies need numpy to compile themselves, but it is not installed yet. I created a conda yaml file instead:

```yaml
name: mixedcoord
channels:
  - conda-forge
dependencies:
  - python=3.9.9
  - ipython=8.0.0
  - matplotlib=3.5.1
  - numpy=1.22.0
  - pandas=1.3.5
  - patsy=0.5.2
  - scikit-image=0.19.1
  - scikit-learn=1.0.2
  - scipy=1.7.3
  - seaborn=0.11.2
  - statsmodels=0.13.1
 ```

I had to upgrade most packages to their last version, as the earlier versions declared in the pipfile did not support osx arm64 architectures. I had apparently no issue when running perform_full_rodrigo.py with best_geerts, but I have not visualized anything yet.

@thomasMisiek Perhaps you can check that these versions still work for you and ship the yaml file with your repo?

```bash
conda env create -f environment.yml
```
