### 2022-01-18 #62 @thomasMisiek
@rougier Sure. My initial work consisted in comparing the biological plausibility of two models of rats' spatial navigation behavior (Geerts et al, 2020, Dollé et al 2010). For this initial project, we decided to use Geerts' original code as the basis of our future implementations, as it already provided the implementation of one of the task we wanted to simulate (Pearce et al, 1998) as well as the navigation model itself.

However, while becoming familiar with Geerts and colleagues code, we found several discrepancies between their simulated protocol and the experimental protocol of Pearce et al (1998). After correcting the protocol, we found that Geerts' model and its original parameters were no longer able to replicate the experimental data. Thus, we created and performed a custom parameter optimization process to check if any new set of parameters was able to replicate Pearce et al, (1998) results. We found that there indeed was such a set of parameters, hence the theoretical model of Geerts was validated even if we initially found that the implementation was faulty.

We argue that the simulated protocol was sufficiently modified for our article to be considered a **replication** of the phenomena observed in Pearce's article rather than a simple reproduction of Geerts' results. This is supported by the fact that Geerts model was initially not able to perform well on the new protocol. Nevertheless, we also provide in the article a separate and successful **reproduction** of Geerts' results with the non-modified protocol.

Finally we can argue that our modified code is also less computationally expensive than the original one, with 25min against 4h30 for 100 agents simulation on Pearce experiment (2 days was only for the parameters optimization process, which we published the results of in the paper). We can also argue that our code is documented, with a functional directory-tree and original notebooks that allows the user to easily visualize extensive results. Moreover, we provide supplementary experiments and spatial cognition models in our implementation, since our Github repository is shared between two related articles. These new models and tasks required refactorisation of the code, and thus brought additional **technical** modifications, which further plead for a replication and not a reproduction according to the ReScience FAQ. Overall these enhancements and modifications prompted us to publish our code, which we believe is an interesting additional option to simulate Geerts' model, beside the ModelDB's one.
