### 2022-01-19 #62 @MehdiKhamassi
In addition to what @thomasMisiek described, when he writes "Nevertheless, we also provide in the article a separate and successful reproduction of Geerts' results with the non-modified protocol.", we can add that even for this reproduction, we had to search for the correct model parameters. This is because the value of a few model parameters were not specified in the paper, and a few other parameters were indicated as having different values than what we found in the original code. So @thomasMisiek had to perform a parameter search to do this reproduction.
We have included a table in our manuscript which specifies what are the correct parameter values for each results figure.
