### 2022-01-19 #62 @thomasMisiek
I added the environment.yml file to the repository, added a few lines about it in the README. I kept the requirements.txt file as another option but updated it.
