### 2022-02-15 #62 @mstimberg
@thomasMisiek I started to work on the review and was about to run the code, but just noticed that the `GEERTS-hippocampus-dls-modified` directory which is supposed to contain the code for the reproduction  is empty. I think it is supposed to be a git submodule, but it does not seem to be pointing anywhere.
