### 2022-02-26 #62 @vitay
@thomasMisiek I finally ran the notebook `Geerts Arbitrator - Egocentric MF-SR.ipynb` (15 hours for the two grid searches...). Unfortunately, when running the cell with:

```python
res_df = plot_all_perfs(path, size_plot=10, relative=True)
```

I got the following error:

```
KeyError                                  Traceback (most recent call last)
File ~/Applications/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/core/indexes/base.py:3361, in Index.get_loc(self, key, method, tolerance)
   [3360](file:///~/Applications/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/core/indexes/base.py?line=3359)[ try:
-> ]()[3361](file:///~/Applications/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/core/indexes/base.py?line=3360)[     return self._engine.get_loc(casted_key)
   ]()[3362](file:///~/Applications/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/core/indexes/base.py?line=3361)[ except KeyError as err:

File ~/Applications/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/_libs/index.pyx:76, in pandas._libs.index.IndexEngine.get_loc()

File ~/Applications/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/_libs/index.pyx:108, in pandas._libs.index.IndexEngine.get_loc()

File pandas/_libs/hashtable_class_helper.pxi:5198, in pandas._libs.hashtable.PyObjectHashTable.get_item()

File pandas/_libs/hashtable_class_helper.pxi:5206, in pandas._libs.hashtable.PyObjectHashTable.get_item()

KeyError: 'se_cluster_100_rodrigo_mean_relative'

KeyError                                  Traceback (most recent call last)
/Users/vitay/Articles/reviews/ReScience-Misiek/sources/notebooks/Geerts Arbitrator - Egocentric MF-SR.ipynb Cell 19' in <module>
----> ]()[1](vscode-notebook-cell:/Users/vitay/Articles/reviews/ReScience-Misiek/sources/notebooks/Geerts%20Arbitrator%20-%20Egocentric%20MF-SR.ipynb#ch0000018?line=0)[ res_df = plot_all_perfs(path, size_plot=10, relative=True)

File ~/Articles/reviews/ReScience-Misiek/sources/grid_search.py:939, in plot_all_perfs(directory, size_plot, relative, mode)
    ]()[936](file:///~/Articles/reviews/ReScience-Misiek/sources/grid_search.py?line=935)[ df_pearce,data1,bin12,bin11 = plot_single_perfs(res_df, size_plot, "inv_temp", "srlr", "pearce", relative)
    ]()[937](file:///~/Articles/reviews/ReScience-Misiek/sources/grid_search.py?line=936)[ _,data4,bin42,bin41 = plot_single_perfs(res_df, size_plot, "qlr", "gamma", "pearce", relative)
--> ]()[939](file:///~/Articles/reviews/ReScience-Misiek/sources/grid_search.py?line=938)[ df_rodrigo,data2,bin22,bin21 = plot_single_perfs(res_df, size_plot, "inv_temp", "srlr", "rodrigo", relative)
    ]()[940](file:///~/Articles/reviews/ReScience-Misiek/sources/grid_search.py?line=939)[ _,data5,bin52,bin51 = plot_single_perfs(res_df, size_plot, "qlr", "gamma", "rodrigo", relative)
    ]()[942](file:///~/Articles/reviews/ReScience-Misiek/sources/grid_search.py?line=941)[ #_,data3,bin32,bin31 = plot_single_perfs(res_df, size_plot, "inv_temp", "srlr", "product", relative)
    ]()[943](file:///~/Articles/reviews/ReScience-Misiek/sources/grid_search.py?line=942)[ #_,data6,bin62,bin61 = plot_single_perfs(res_df, size_plot, "qlr", "gamma", "product", relative)

File ~/Articles/reviews/ReScience-Misiek/sources/grid_search.py:806, in plot_single_perfs(df, bins_nbr, dim1_name, dim2_name, expe, relative)
    ]()[804](file:///~/Articles/reviews/ReScience-Misiek/sources/grid_search.py?line=803)[ elif expe == "rodrigo":
    ]()[805](file:///~/Articles/reviews/ReScience-Misiek/sources/grid_search.py?line=804)[     if relative:
--> ]()[806](file:///~/Articles/reviews/ReScience-Misiek/sources/grid_search.py?line=805)[         data[dim1][dim2] = df["se_cluster_100_rodrigo_mean_relative"][indexes].min()
    ]()[807](file:///~/Articles/reviews/ReScience-Misiek/sources/grid_search.py?line=806)[     else:
    ]()[808](file:///~/Articles/reviews/ReScience-Misiek/sources/grid_search.py?line=807)[         data[dim1][dim2] = df["se_cluster_100_rodrigo_mean_absolute"][indexes].min()

File ~/Applications/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/core/frame.py:3458, in DataFrame.__getitem__(self, key)
   ]()[3456](file:///~/Applications/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/core/frame.py?line=3455)[ if self.columns.nlevels > 1:
   ]()[3457](file:///~/Applications/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/core/frame.py?line=3456)[     return self._getitem_multilevel(key)
-> ]()[3458](file:///~/Applications/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/core/frame.py?line=3457)[ indexer = self.columns.get_loc(key)
   ]()[3459](file:///~/Applications/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/core/frame.py?line=3458)[ if is_integer(indexer):
   ]()[3460](file:///~/Applications/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/core/frame.py?line=3459)[     indexer = [indexer]

File ~/Applications/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/core/indexes/base.py:3363, in Index.get_loc(self, key, method, tolerance)
   ]()[3361](file:///~/Applications/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/core/indexes/base.py?line=3360)[         return self._engine.get_loc(casted_key)
   ]()[3362](file:///~/Applications/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/core/indexes/base.py?line=3361)[     except KeyError as err:
-> ]()[3363](file:///~/Applications/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/core/indexes/base.py?line=3362)[         raise KeyError(key) from err
   ]()[3365](file:///~/Applications/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/core/indexes/base.py?line=3364)[ if is_scalar(key) and isna(key) and not self.hasnans:
   ]()[3366](file:///~/Applications/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/core/indexes/base.py?line=3365)[     raise KeyError(key)

KeyError: 'se_cluster_100_rodrigo_mean_relative']()
```

Is there a way to fix this without losing all the simulations?
