### 2022-02-26 #62 @thomasMisiek
@vitay Yes, setting the "relative" parameter to True in the "compute_clusters_perfs(path, 100, experimental_data_pearce, experimental_data_rodrigo, relative=False)" cell, and running again the "res_df = plot_all_perfs(path, size_plot=10, relative=True)" cell should do the trick. Sorry for this inconvenience. I will update the repository.
