### 2022-02-27 #62 @vitay
@thomasMisiek Thanks, that allowed me to continue running the notebook. I attach the pdf of the notebook. It seems that the Rodrigo experiments quite failed, there was no better parameter values. I am a bit lost in all the other plots. Can you check they are what you intended?

[Geerts Arbitrator - Egocentric MF-SR.pdf](https://github.com/ReScience/submissions/files/8149301/Geerts.Arbitrator.-.Egocentric.MF-SR.pdf)
