### 2022-02-28 thomasMisiek/mixed-coordination-models#1 @mstimberg
Improve installation instructions for Geerts et al. reproduction
[Opened as part of the ReScience review: ReScience/submissions#62]

The README for the Geerts et al. reproduction in `GEERTS-hippocampus-dls-modified` does not have any installation instructions, but running it with only the requirements from the main repository fails because the `tqdm` and `cython` dependencies are missing.

There is also a `setup.py` file (which declares `tqdm`as a dependency among other things, but requires Cython to be run), but I'm not 100% sure of how it is supposed to be used. Running it with `python setup.py build_ext` for example fails with: `'GEERTS-hippocampus-dls-modified.hippocampus.fastBVC' is not a valid module name`. Maybe it is only necessary for the `pyximport` in the code? Might be good to mention this in the README.

As a bit of an aside, I'm not sure how useful a `run.ipynb` with a single `import pearce` line is...
