### 2022-02-28 thomasMisiek/mixed-coordination-models#2 @mstimberg
Script `pearce.py` for Geerts et al. fails with seaborn/pandas error
[Opened as part of the ReScience review: ReScience/submissions#62]

After fixing the dependencies, running `pearce.py` works and prints something (even though it is quite unclear what the meaning of that print output is). In the end, however, it fails with an error during the plotting procedure:
```Python
Traceback (most recent call last):
  File "/home/marcel/programming/mixed-coordination-models/GEERTS-hippocampus-dls-modified/pearce.py", line 159, in <module>
    ax = sns.lineplot(data=first_and_last, x='session', y='escape time', hue='trial')
  File "/mnt/data/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/seaborn/_decorators.py", line 46, in inner_f
    return f(**kwargs)
  File "/mnt/data/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/seaborn/relational.py", line 710, in lineplot
    p.plot(ax, kwargs)
  File "/mnt/data/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/seaborn/relational.py", line 471, in plot
    for sub_vars, sub_data in self.iter_data(grouping_vars, from_comp_data=True):
  File "/mnt/data/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/seaborn/_core.py", line 983, in iter_data
    data = self.comp_data
  File "/mnt/data/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/seaborn/_core.py", line 1057, in comp_data
    comp_col.loc[orig.index] = pd.to_numeric(axis.convert_units(orig))
  File "/mnt/data/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/core/indexing.py", line 723, in __setitem__
    iloc._setitem_with_indexer(indexer, value, self.name)
  File "/mnt/data/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/core/indexing.py", line 1732, in _setitem_with_indexer
    self._setitem_single_block(indexer, value, name)
  File "/mnt/data/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/core/indexing.py", line 1959, in _setitem_single_block
    value = self._align_series(indexer, Series(value))
  File "/mnt/data/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/core/indexing.py", line 2096, in _align_series
    ser = ser.reindex(obj.axes[0][indexer[0]], copy=True)._values
  File "/mnt/data/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/core/series.py", line 4580, in reindex
    return super().reindex(index=index, **kwargs)
  File "/mnt/data/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/core/generic.py", line 4818, in reindex
    return self._reindex_axes(
  File "/mnt/data/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/core/generic.py", line 4839, in _reindex_axes
    obj = obj._reindex_with_indexers(
  File "/mnt/data/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/core/generic.py", line 4883, in _reindex_with_indexers
    new_data = new_data.reindex_indexer(
  File "/mnt/data/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/core/internals/managers.py", line 670, in reindex_indexer
    self.axes[axis]._validate_can_reindex(indexer)
  File "/mnt/data/miniforge3/envs/mixedcoord/lib/python3.9/site-packages/pandas/core/indexes/base.py", line 3785, in _validate_can_reindex
    raise ValueError("cannot reindex from a duplicate axis")
ValueError: cannot reindex from a duplicate axis
```
Seaborn 0.11.2, pandas 1.3.5
