### 2022-02-28 thomasMisiek/mixed-coordination-models#3 @mstimberg
Clarify "custom" parameter set
[Opened as part of the ReScience review: ReScience/submissions#62]

The README states:
> "'custom' is a random set of parameters that is intended to be modified by the user to explore new models behaviors on the task."

I think this could need a bit of clarification, in particular *how* it is supposed to be modified (presumably by editing the corresponding lines in the script).
