### 2022-02-28 thomasMisiek/mixed-coordination-models#4 @mstimberg
case mismatch in file names for `perform_full_pearce.py`
[Opened as part of the ReScience review: ReScience/submissions#62]

At the end of a `perform_full_pearce.py` run, I get an error
```
FileNotFoundError: [Errno 2] No such file or directory: '../images/results_pearce.jpg'
```
The reason is that a number of files are named `.JPG` instead of `.jpg` (and I'm on Linux with a case-sensitive file system).
