### 2022-02-28 thomasMisiek/mixed-coordination-models#5 @mstimberg
Parameter mismatch in `perform_full_pearce` and `perform_full_exp3_pearce`
[Opened as part of the ReScience review: ReScience/submissions#62]

When I run the 3rd Pearce experiment after the 1st one, I get the error:
```
Error: Experience 3 of Pearce 1998 can only be conducted with agents trained on the first experiment
Please train rats on the first experiment of Pearce using perform_full_pearce script,
then execute perform_full_exp_pearce using the same parameters and directory

command example:
'python perform_full_pearce.py mydirectory best_geerts'
'python perform_full_exp3_pearce.py mydirectory best_geerts'

check that the first experiment and third experiment parameters are identical, if an error persist
```
As the last line of the error suggests, this is due to a parameter mismatch (value for `n_agents` for `best_geerts` is 20 vs. 100 in both files).
