### 2022-02-28 thomasMisiek/mixed-coordination-models#6 @mstimberg
Error when running `perform_full_exp3_pearce.py`
After fixing the parameter mismatch mentioned in #5 , the script fails with the following error:
```pytb
Traceback (most recent call last):
  File "/home/marcel/programming/mixed-coordination-models/sources/perform_full_exp3_pearce.py", line 142, in <module>
    perform_full_exp3_pearce()
  File "/home/marcel/programming/mixed-coordination-models/sources/perform_full_exp3_pearce.py", line 131, in perform_full_exp3_pearce
    perform_group_exp3_pearce(env_params, ag_params, directory=sys.argv[1]+"/control_group", show_plots=False, save_plots=True)
  File "/home/marcel/programming/mixed-coordination-models/sources/exp3_pearce_protocol.py", line 112, in perform_group_exp3_pearce
    res2_mf, res2_allo, res2_sr, res2_combined = get_mean_preferred_dirs(agents, 0, 0)
ValueError: too many values to unpack (expected 4)
```
