### 2022-02-28 #62 @mstimberg
Hi everyone. I also started running the code, but began from a different angle than @vitay . I had a few small-ish problems, and opened corresponding issues in the code repository (thomasMisiek/mixed-coordination-models#1, thomasMisiek/mixed-coordination-models#2, thomasMisiek/mixed-coordination-models#3, thomasMisiek/mixed-coordination-models#4, thomasMisiek/mixed-coordination-models#5).
