### 2022-02-28 #62 @thomasMisiek
@vitay I just checked your PDF and it looks like what I intended, with the exception of the display of rodrigo's grid-search results. I believe the rodrigo's grid-search and display methods themselves are working properly. However, nothing is displayed because datapoints are only plotted if they are associated to behavioral performances which validates a set of statistical tests. On my last execution of the code, I placed very low values of p in the source code such that only 8 sets of parameters out of 2000 validated the statistical tests. It seems that in your execution, no datapoint at all were able to validate these tests. Increasing p-values from 0.005 to 0.05, for example, should allow much more datapoints to be displayed, as with Pearce's grid-search.

I also see in your PDF that the plot_all_perfs() function did not display the three sets of parameters producing the lowest MSE, as it should have, and only printed "<IPython.core.display.HTML object>". This never happened on my machines, I will try to find a more robust and portable way to plot this HTML object on any computer.

Other than that, every cells that you executed produced the intended results for Pearce experiment as well as for Rodrigo's experiment. Nevertheless, you should ignore every results that comes after the markdown cell **2.3 Results on Pearce’s experiment with steeper decrease in P(SR)**. Nothing after this cell is related to our replication study, but is related to another paper. This includes Rodrigo's experiment (grey histograms) and Pearce's 3rd experiment (blue histograms).

Beside the grid-search, the only part related to this replication study is the **(2. Testing the grid-search predictions: Pearce, Roberts, Good)** section. The best set of parameters found using the grid-search is defined **(2.1)**, then 100 agents simulations are performed for the control group **(2.1)** and HPC-lesioned groups **(2.2)**. The escape time of both groups are then plotted on the same figure with the **plot_pearce()** function. There is a lot of information displayed in the **(2.x)** section, the only relevant informations for the replication paper are the "Our results" plot in the **plot_pearce()** cell just before the **(2.3)** section. And also the ANOVAs on the "trial" variable for the control group**(2.1)** and lesioned group **(2.2)**, and the ANOVAs on the "group" variable in the **plot_pearce()** cell. Your results (ANOVAs and figure) seems to be coherent with those in our papers.
