### 2022-03-01 #62 @vitay
@thomasMisiek Thanks. Would it be possible to add a simpler notebook that only deals with the results presented in the ReScience paper? It is nice to have code with a broader scope, but some readers might want to focus on Geerts' model applied to Pearce's task.
