### 2022-03-03 #62 @mstimberg
I concur with @vitay's suggestion, in particular given that these simulations take a lot of time to run. I tried running the notebook overnight on my notebook, but it failed due to a disk space issue when the results folder was filled with several GB. Could you give an estimate how much disk space is needed?
An unrelated question: there are a number of lines in the notebook annotated with comments like `# lower values of p`, but I don't see anything in the function calls that changes `p`. The most confusing are these three repeated lines:
```Python
res_df = plot_all_perfs(path, size_plot=10, relative=True) # higher values of p, compared to the next cell
res_df = plot_all_perfs(path, size_plot=10, relative=True) # lower values of p
res_df = plot_all_perfs(path, size_plot=10, relative=True) # lower values of p
```
These surely all do the same thing?
