### 2022-03-03 #62 @thomasMisiek
@vitay @mstimberg Ok, I will create a specific notebook for the replication study right now, with only cells related to Pearce experiment. Please note that out of all the notebooks, only the "Geerts Arbitrator - Egocentric MF-SR.ipynb" is currently related to this paper. I understand that it might have not been clear from the beginning, thus I will add a few lines in the README to specify the required steps for people that are only interested in this replication paper.

@mstimberg On my computer, a complete grid-search simulation for Pearce experiment generates 10 Go of data. A 100 datapoints experiment generates 400 Mo of data. It's respectively 11Go and 1Go for Rodrigo's task, which won't appear in the new notebook anyway.  Running Pearce's grid-search should take 4 to 6 hours, whereas Rodrigo's should take 10 to 12 hours. The remaining cells of the notebook should fully execute in less than one hour.

There is indeed three cells plotting the grid-search results with the same function call `plot_all_perfs(path, size_plot=10, relative=True).` If you open the notebook without modifications, you will see that they display different results. That is because after each function call, I modified the list of p-values directly in the code. The reason for this was to show the user that modifying p-values have an effect on the grid-search results. I understand that this can be misleading for the user after running again the whole notebook, I will thus remove two of the three cells and add a comment to specify that the p-values can be found and modified in the grid-search.py module.
