### 2022-03-03 #62 @thomasMisiek
@mstimberg @vitay I just added the Geerts Replication.ipynb to the repository, it contains everything related to the paper (replication) and nothing else. I added a paragraph about it in the README. I also removed two of the three displays of the grid-search.
@mstimberg I will solve the opened issue 1 and 2 as soon as possible, on saturday.
