### 2022-03-11 thomasMisiek/mixed-coordination-models#7 @mstimberg
`random_grid_search` fails if `results` dir does not exist
[Opened as part of the ReScience review: https://github.com/ReScience/submissions/issues/62]

Running e.g. the `Geerts Replication.ipynb` notebook in a freshly cloned repository will fail, since it assumes that the `results` directory exists. Using `os.makedirs` with `exists_ok=True` could simplify things here.
