### 2022-03-11 thomasMisiek/mixed-coordination-models#8 @mstimberg
Estimated time in `random_grid_search` seems far off
[Opened as part of the ReScience review: https://github.com/ReScience/submissions/issues/62]

The random grid search tries to estimate the remaining time, but the estimate seems far off: it always seems to display an estimated remaining time of about 1h, but the total grid search took ~10h. It might be that at later stages of the grid search the estimates become more reliable, since I only actively watched a few iterations at the beginning.
