### 2022-03-11 #62 @mstimberg
I successfully ran the `Geerts Replication.ipynb` notebook and only encountered two small issues (thomasMisiek/mixed-coordination-models#7, thomasMisiek/mixed-coordination-models#8). The results look comparably to the values in the unchanged notebook, but not identical. I assume part of it is due to randomness in the simulations, would it be possible to fix things with a seed to get the exact same results? Regardless of this, please re-run/save the notebook. Its cells were evaluated out-of-order, and it looks as if the grid search was manually restricted to 150 instead of 2000 data points for the earlier part of the notebook, while it uses the results from a full grid search in the later parts. In the "Lesioned agents training" section, `perform_group_pearce` (or a function called by it) seems to be different from the version in the repository: in the uploaded notebook it performs a "Two-way ANOVA on trial and session" whereas when I run it, it performs two separate ANOVAs on trial and session.

> The reason for this was to show the user that modifying p-values have an effect on the grid-search results. I understand that this can be misleading for the user after running again the whole notebook, I will thus remove two of the three cells and add a comment to specify that the p-values can be found and modified in the grid-search.py module.

If the p-value cut-off is an interesting parameter to vary here, it could be added as an optional parameter to the function, maybe?
