### 2022-03-18 thomasMisiek/mixed-coordination-models#2 @thomasMisiek
I just fixed it.
The plotting failed because of seaborn 0.11.2. It now works using seaborn 0.9.0 which is set in the requirements.txt and environment.yml file.
The nonsensical printing during the execution is replaced with a stable tqdm display of the evolution of the experiment (current agent - current session - current trial).
