### 2022-03-18 thomasMisiek/mixed-coordination-models#1 @thomasMisiek
@mstimberg Ok, sorry for this inconvenience. I added both a requirement.txt and a environment.yml file. in the `GEERTS-reproduction` (former `GEERTS-hippocampus-dls-modified`) repository.
The instructions on how to use the repository are now written in the README.md of `GEERTS-reproduction`.
No need to use or execute the setup.py file.
And finally I modified the run.ipynb notebook to display more useful information.

Note that this reproduction repository is mostly identical to what is accessible on Geerts' ModelDB repository. I only modified the models' parameters in order to reproduce their results (all modifications are now detailed in the README).
