### 2022-03-19 #62 @thomasMisiek
@mstimberg I added a seed parameter to both the `random_grid-search()` function and the `perform_group_pearce()` function.

I am re-running the notebook again to get rid of the out-of-order cells calls, the 150 datapoints instead of 2000, and the "Two-way ANOVA on trial and session" which comes from a previous version of the program. I will also run two successive grid-search with 2000 datapoints. The first being identical to the current one in the notebook, and the second exploring only a subset of the multidimensional space from the first, to improve the accuracy of the search. I will inform you when this notebook is ready and uploaded.

I added a single p-value parameter to the function that performs the statistical analyses, so that the user can tune it.
