### 2022-03-19 thomasMisiek/mixed-coordination-models#8 @thomasMisiek
The format of the estimated time was day:hour:minute:second, and there was always a day to substract from the counter. I modified the function to print a more explicit format and not to display this additional day.
