### 2022-03-20 #62 @thomasMisiek
I have modified the PDF (parameters values and ANOVAs' results) to fit the new results of the geerts_replication notebook. I also created a "best_geerts" and "best_geerts_stochastic" option, for respectively deterministic and stochastic computing if the user chooses to use the command line instead of the notebook.
