### 2022-03-23 #62 @vitay
Thanks for the changes and the new notebook. The code runs without major problems (and much faster) and I got the following results for the deterministic case:

Just a tip: add `#cython: language_level=3` at the top of each of your .pyx files to get rid of the warning:

```
FutureWarning: Cython directive 'language_level' not set, using 2 for now (Py2). This will change in a later release!
```

The tqdm progress bar did not work for me, but it is not a big deal, it probably comes from my setup.

I guess it is time to conclude the review now. Overall this is a very valuable partial replication of Geerts' model, as it fixes several issues with the published code and gets even closer to Pearce's experimental data. The goals of the replication are very well explained in the paper and the code will be useful to every researchers interested in spatial navigation and successor representations.

@rougier I recommend accepting the paper as a partial replication.
