### 2022-03-30 thomasMisiek/mixed-coordination-models#6 @mstimberg
This still fails for me with the same error message. But running this experiment is not necessary fo ReScience/submissions#62 if I am not mistaken. Actually, I am not quite sure what this "third experiment" corresponds to, as far as I can tell, [Pearce et al. 1998](https://doi.org/10.1038/23941) only has two experiments, and I don't see anything in the article about inverting the landmark to platform distance (as mentioned in the function's docstring).
