### 2022-03-30 #62 @mstimberg
Hi @rougier, @thomasMisiek. I am satisfied in general and happy with the various improvements, they should make running the code much more straightforward. I have one issue open in the repository that is still not addressed (thomasMisiek/mixed-coordination-models#6), but it might not be relevant for the ReScience paper (I'll let @thomasMisiek confirm). Apart from that, I was able to run everything and get the expected results, both with the notebooks and the scripts.
I therefore also recommend accepting the paper. Please note that the generated PDF still has a number of formatting issues, though: The article reference at the start is not formatted correctly in the usual "block", and the sections seem to be "on the wrong level" (e.g. `\subsection` instead of `section`), leading to the 0.1, 0.2, ... numbering. It also states "Received: 01 November 2018", but I assume this gets fixed when the other missing metadata will be added. As a final note: make sure to tag a new release and publish it on Zenodo, the PDF currently still links to https://zenodo.org/record/5854457 with the version from January.
