### 2022-03-30 #62 @mstimberg
Apologies, I forgot one minor suggestion in my previous comment: you might consider adding a footnote to the paper explaining your understanding of "replication" and "reproduction" (or simply link to the [ReScience FAQ](https://rescience.github.io/faq/)). Your manuscript is mostly about a replication, but you also briefly mention the reproduction, and there is a lot of confusion around these terms.
