### 2022-03-30 thomasMisiek/mixed-coordination-models#6 @thomasMisiek
Ah sorry, I did not notice this #6th issue.

Yes, this experiment is not related to ReScience/submissions#62, and everything related to it has been removed from the "Geerts Replication" notebook.

However. I acknowledge that the name "Third experiment" is misleading, and I will find a better name for it in our future article. Pearce et al. 1998 explicitly mentioned two experiments. First, the experiment with HPC-lesioned and control groups. And additionally, the experiment with "trials" and "sessions" groups. What I call the third experiment is the supplementary trial added at the end of the first experiment, where the landmark is placed at the center of the maze, and with an inverted platform direction relative to the landmark. Citing Pearce's paper:

> For half of the rats in each group the platform was located at the usual direction and distance from the landmark, and for the remaining rats the platform was located at the same distance from the landmark but on the opposite side to normal.

In Geerts code, the landmark-platform distance encodes how far north of the platform the landmark is placed. Negative distance values switch this direction south. This being said, I might replace "inverting the landmark to platform distance" in the docstring to "inverting the landmark to platform direction" if that can make things clearer.

Furthermore I noticed that it is the landmark that should be placed at the center of the maze and not the platform like in my design, I will modify this.

And of course I will update the `res2_mf, res2_allo, res2_sr, res2_combined = get_mean_preferred_dirs(agents, 0, 0)` to remove the bug you described in your comment.
