### 2022-03-30 #62 @thomasMisiek
@vitay @mstimberg  Thank you for your thoughtful comments and your efforts towards improving the code and the readability of the results.

@rougier @mstimberg Yes I confirm that [https://github.com/thomasMisiek/mixed-coordination-models/issues/6](url) is not relevant for the Rescience paper.

@mstimberg I will add a footnote to the PDF about Replication/Reproduction and then I will fix the PDF format issues you mentioned, as well as the Zenodo issue.
