### 2022-03-31 #62 @rougier
@thomasMisiek Congratulations, your paper is accepted! @vitay @mstimberg Many thanks again for efficient and useful review.

@thomasMisiek I'll proceed with publication which can take a few days, don't hesitate to remind/spam me if it's not done in the next few days.
