### 2022-05-19 #62 @rougier
For publication, I would need a software heritage id. Can you check https://www.softwareheritage.org/save-and-reference-research-software/ and get sch id (it's quite easy to generate).
