### 2022-05-22 #62 @thomasMisiek
@rougier Here it is https://archive.softwareheritage.org/browse/revision/239cae046e09974d285fbb0d06c734b3a901fb5c/?origin_url=https://github.com/thomasMisiek/mixed-coordination-models
